[HL25] Backport (but disabled) Steam Rich Presence commands call
The primary reason for disabling this is because these console commands
have no effect when running a mod (started through Steam at least, not
tested when launched directly from command-line).

Mods that are hosted on Steamworks might be able to make use of those
by doing the necessary back-end setup and compiling their client project
with the "STEAM_RICH_PRESENCE" pre-processor define.

Co-Authored-By: Joël Troch <JoelTroch@users.noreply.github.com>